@@ -1,26 +1,68 @@
 {
-  "name": "profile",
-  "version": "0.1.0",
-  "private": true,
+  "name": "nextjs-resume",
+  "version": "5.1.1",
+  "license": "MIT",
+  "type": "module",
+  "author": {
+    "name": "Colin Hemphill",
+    "email": "colin@colinhemphill.com",
+    "url": "https://colinhemphill.com"
+  },
+  "keywords": [
+    "resume",
+    "generator",
+    "nextjs",
+    "pdf",
+    "tailwind"
+  ],
+  "homepage": "https://github.com/colinhemphill/nextjs-resume",
   "scripts": {
-    "dev": "next dev",
     "build": "next build",
+    "build:content": "contentlayer build",
+    "dev": "next dev",
+    "format": "prettier --write 'src/**/*.{json,js,ts,tsx}'",
+    "lint": "next lint",
+    "prepare": "husky",
     "start": "next start",
-    "export": "next export",
-    "lint": "next lint"
+    "test": "bun test",
+    "test:coverage": "bun test --coverage",
+    "test:update": "bun test --update-snapshots",
+    "test:watch": "bun test --watch",
+    "types": "tsc -p tsconfig.json --noEmit"
   },
   "dependencies": {
-    "@types/node": "20.3.1",
-    "@types/react": "18.2.13",
-    "@types/react-dom": "18.2.6",
-    "autoprefixer": "10.4.14",
-    "eslint": "8.43.0",
-    "eslint-config-next": "13.4.7",
-    "next": "13.4.7",
-    "postcss": "8.4.24",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "3.3.2",
-    "typescript": "5.1.3"
+    "@heroicons/react": "2.1.5",
+    "@icons-pack/react-simple-icons": "9.6.0",
+    "@radix-ui/colors": "3.0.0",
+    "@react-pdf/renderer": "3.4.4",
+    "class-variance-authority": "0.7.0",
+    "contentlayer": "0.3.4",
+    "next": "14.2.5",
+    "next-contentlayer": "0.3.4",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-pdf-html": "2.1.2",
+    "tailwind-merge": "2.5.1"
+  },
+  "devDependencies": {
+    "@happy-dom/global-registrator": "14.12.3",
+    "@testing-library/react": "16.0.0",
+    "@types/node": "22.2.0",
+    "@types/react": "18.3.3",
+    "autoprefixer": "10.4.20",
+    "bun-types": "1.1.22",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.5",
+    "eslint-config-prettier": "9.1.0",
+    "husky": "9.1.4",
+    "jsdom": "24.1.1",
+    "lint-staged": "15.2.8",
+    "postcss": "8.4.41",
+    "prettier": "3.3.3",
+    "prettier-plugin-tailwindcss": "0.6.6",
+    "tailwindcss": "3.4.9",
+    "typescript": "5.5.4",
+    "webpack": "5.93.0",
+    "windy-radix-palette": "2.0.0-beta.7"
   }
 }
