@@ -1,21 +1,88 @@
-import './globals.css'
-import { Inter } from 'next/font/google'
+import * as colors from '@radix-ui/colors';
+import { Metadata, Viewport } from 'next';
+import { Albert_Sans, JetBrains_Mono } from 'next/font/google';
+import { PropsWithChildren } from 'react';
+import resumeConfig from '../../edit-me/config/resumeConfig';
 
-const inter = Inter({ subsets: ['latin'] })
+// STYLES
+import { personal } from '@content';
+import { headers } from 'next/headers';
+import { protocol, vercelURL } from 'src/helpers/env';
+import { fullName } from 'src/helpers/utils';
+import { twMerge } from 'tailwind-merge';
+import { ThemeSetting } from '../../edit-me/types/Config';
+import './globals.css';
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+const accentColor = resumeConfig.accentColor;
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+const albert = Albert_Sans({
+  display: 'swap',
+  subsets: ['latin'],
+  variable: '--font-albert',
+});
+
+const jetBrainsMono = JetBrains_Mono({
+  display: 'swap',
+  subsets: ['latin'],
+  variable: '--font-jetbrains-mono',
+});
+
+export const generateMetadata = async (): Promise<Metadata> => {
+  const host = headers().get('host');
+  const baseURL = `${protocol}://${host || vercelURL}`;
+  const siteName = `${fullName} Professional Résumé`;
+  const title = `Résumé | ${fullName} | Somewhere`;
+  const description = `Professional résumé for ${fullName}.`;
+
+  return {
+    metadataBase: new URL(baseURL),
+    applicationName: siteName,
+    authors: { name: fullName },
+    creator: fullName,
+    description,
+    generator: 'Next.js',
+    keywords: ['resume', fullName, 'next.js', 'pdf'],
+    openGraph: {
+      type: 'profile',
+      firstName: personal.givenName,
+      lastName: personal.familyName,
+      title,
+      description,
+      siteName,
+      url: baseURL,
+    },
+    title,
+    twitter: {
+      site: siteName,
+      creator: fullName,
+      description,
+      title,
+    },
+  };
+};
+
+export const viewport: Viewport = {
+  initialScale: 1,
+  // @ts-ignore
+  themeColor: colors[accentColor][`${accentColor}9`],
+  width: 'device-width',
+};
+
+const RootLayout: React.FC<PropsWithChildren> = async ({ children }) => {
   return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
+    <html
+      lang="en"
+      className={twMerge(
+        albert.variable,
+        jetBrainsMono.variable,
+        resumeConfig.appTheme === ThemeSetting.Dark && 'dark',
+      )}
+    >
+      <body className="bg-neutral-1 text-neutral-12 selection:bg-accent-11 selection:text-neutral-1">
+        {children}
+      </body>
     </html>
-  )
-}
+  );
+};
+
+export default RootLayout;
